Surveying The Avalanche: What happens to the draft board when there's a run at a position?
Participant: Dan Falkenheim
Notes: I uploaded a .md file which has all of the tables and figures embedded and a .rmd file as well. Let me know if you want me to upload a PDF, and if you need me to cut down on the amount of figures. I counted each facet grid as one figure, so I wasn't sure if there were guidelines on that and if I accidentally went over.